refactor(config): extract configService — move state + logic out of route file
getConfig/getDriverOfDay/generateOrderId/isPastCutoff/getActiveSeasonalCategory
and all config state (DEFAULTS, loadConfig, migrations, setInterval reminder)
move to services/configService.js. settings.js becomes a thin HTTP layer.

Re-exports from settings.js for backward compat — callers can migrate
to importing from configService directly over time.
