+++ backend/src/routes/settings.js
@@ -1,363 +1,28 @@
-// Settings routes — config persisted to Airtable, daily settings in memory.
-// Like a factory config binder stored in the central filing cabinet (Airtable)
-// instead of on a whiteboard that gets erased when the lights go off.
-// Daily preferences (driver-of-day) still reset at midnight — that's intentional.
-
+// Settings routes — HTTP layer for app configuration.
+// State + business logic lives in services/configService.js.
 import { Router } from 'express';
 import { authorize } from '../middleware/auth.js';
 import { getBackupDriverName, setBackupDriverName } from '../services/driverState.js';
-import * as db from '../services/airtable.js';
-import { TABLES } from '../config/airtable.js';
-import { sendAlert } from '../services/telegram.js';
 import { DELIVERY_STATUS } from '../constants/statuses.js';
-import * as appConfigRepo from '../repos/appConfigRepo.js';
-import * as productConfigRepo from '../repos/productConfigRepo.js';
 import * as orderRepo from '../repos/orderRepo.js';
+import {
+  getConfig, getAllConfig, updateConfigBulk,
+  getDriverOfDay, setDailyDriver, getDailyState,
+  driverNames, autoClearIfNewDay,
+} from '../services/configService.js';
+
+// Re-export for backward-compat with existing callers that haven't been updated yet.
+// TODO: remove once all callers import directly from configService.
+export { getConfig, updateConfig, generateOrderId, isPastCutoff, getActiveSeasonalCategory, getDriverOfDay } from '../services/configService.js';
 
 const router = Router();
-
-// ── Default configuration values ────────────────────────────
-// Used as fallback if Airtable config row doesn't exist yet or is empty.
-const DEFAULTS = {
-  defaultDeliveryFee: 35,
-  targetMarkup:       2.2,
-  suppliers:          ['Stojek', '4f', 'Stefan', 'Mateusz', 'Other'],
-  stockCategories:    ['Roses', 'Tulips', 'Seasonal', 'Greenery', 'Accessories', 'Other'],
-  paymentMethods:     ['Cash', 'Card', 'Mbank', 'Monobank', 'Revolut', 'PayPal', 'Wix Online'],
-  orderSources:       ['In-store', 'Instagram', 'WhatsApp', 'Telegram', 'Wix', 'Flowwow', 'Other'],
-  driverCostPerDelivery: 35,
-  driverCostPerPORun: 45,
-  floristNames: ['Anya', 'Daria'],
-  rateTypes: ['Standard', 'Wedding', 'Holidays'],
-  floristRates: {},
-  extraDrivers: [],
-  storefrontCategories: {
-    permanent: [
-      { name: 'All Bouquets', slug: 'all-bouquets', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-      { name: 'Bestsellers',  slug: 'bestsellers',  description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-    ],
-    seasonal: [
-      { name: "Valentine's Day", slug: 'valentines-day', from: '01-25', to: '02-15', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-      { name: "Women's Day",     slug: 'womens-day',     from: '02-25', to: '03-10', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-      { name: "Mother's Day",    slug: 'mothers-day',    from: '04-20', to: '05-26', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-      { name: 'Easter',          slug: 'easter',         from: '03-28', to: '04-15', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-      { name: 'Christmas',       slug: 'christmas',      from: '12-01', to: '12-26', description: '', translations: { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } } },
-    ],
-    auto: [
-      {
-        name: 'Available Today',
-        slug: 'available-today',
-        description: 'Bouquets ready for same-day delivery or pickup.',
-        translations: {
-          en: { title: 'Available Today',   description: 'Bouquets ready for same-day delivery or pickup.' },
-          pl: { title: 'Dostępne dziś',     description: 'Bukiety gotowe do dostawy lub odbioru tego samego dnia.' },
-          ru: { title: 'Доступно сегодня',  description: 'Букеты, готовые к доставке или самовывозу сегодня.' },
-          uk: { title: 'Доступно сьогодні', description: 'Букети, готові до доставки або самовивозу сьогодні.' },
-        },
-      },
-    ],
-    autoSchedule: true,
-    manualOverride: null,
-    wixCategoryMap: {},
-  },
-  deliveryZones: [
-    { id: 1, name: 'Central Krakow', fee: 35, postcodes: ['30-0', '30-1', '31-0'] },
-    { id: 2, name: 'Suburbs',        fee: 50, postcodes: ['32-0'] },
-    { id: 3, name: 'Out of city',    fee: 80, postcodes: [] },
-  ],
-  freeDeliveryThreshold: 300,
-  expressSurcharge: 20,
-  deliveryTimeSlots: ['10:00-12:00', '12:00-14:00', '14:00-16:00', '16:00-18:00'],
-  availableTodayCutoff: '18:00',
-  availableTodayTimezone: 'Europe/Warsaw',
-  cutoffReminderLastDate: null,
-  slotLeadTimeMinutes: 30,
-  // Shows the per-row "Reconcile premade" action on stock rows in the dashboard.
-  // Off by default — it's admin tooling for fixing historical data mismatches,
-  // not something the florist should tap in daily use.
-  showStockRepairTools: false,
-};
-
-// ── In-memory config (loaded from Postgres on startup) ──────
-let config = structuredClone(DEFAULTS);
-let configLoaded = false;
-
-/**
- * Load config from Postgres App Config table.
- * Merges stored values over defaults so new keys auto-appear.
- */
-async function loadConfig() {
-  try {
-    const stored = await appConfigRepo.get('config');
-    if (stored) {
-      config = deepMerge(DEFAULTS, stored);
-      migrateSeasonalDates();
-      migrateCategoryObjects();
-      migrateAutoCategoryTranslations();
-      migrateFloristRates();
-      console.log('[SETTINGS] Config loaded from Postgres');
-    } else {
-      await appConfigRepo.set('config', DEFAULTS);
-      console.log('[SETTINGS] Config row created in Postgres with defaults');
-    }
-  } catch (err) {
-    console.error('[SETTINGS] Failed to load config from Postgres:', err.message);
-    console.warn('[SETTINGS] Using in-memory defaults');
-  }
-  configLoaded = true;
-}
-
-/**
- * Save current config to Postgres.
- */
-async function saveConfig() {
-  try {
-    await appConfigRepo.set('config', config);
-  } catch (err) {
-    console.error('[SETTINGS] Failed to save config to Postgres:', err.message);
-  }
-}
-
-/**
- * Deep merge: target gets source values, preserving nested structure.
- * Arrays are replaced (not merged) — source array wins entirely.
- */
-function deepMerge(target, source) {
-  const result = { ...target };
-  for (const key of Object.keys(source)) {
-    if (
-      source[key] !== null &&
-      typeof source[key] === 'object' &&
-      !Array.isArray(source[key]) &&
-      typeof target[key] === 'object' &&
-      !Array.isArray(target[key])
-    ) {
-      result[key] = deepMerge(target[key] || {}, source[key]);
-    } else {
-      result[key] = source[key];
-    }
-  }
-  return result;
-}
-
-/**
- * Normalize seasonal dates to MM-DD format.
- * Handles DD-MM, DD.MM, MM.DD — auto-detects by checking if first part > 12.
- */
-function normalizeMMDD(val) {
-  if (!val) return val;
-  const clean = val.replace(/\./g, '-');
-  const parts = clean.split('-');
-  if (parts.length !== 2) return clean;
-  const [a, b] = parts.map(p => p.trim().padStart(2, '0'));
-  if (Number(a) > 12) return `${b}-${a}`;
-  return `${a}-${b}`;
-}
-
-function migrateSeasonalDates() {
-  const sc = config.storefrontCategories;
-  if (!sc?.seasonal) return;
-  let changed = false;
-  for (const entry of sc.seasonal) {
-    const newFrom = normalizeMMDD(entry.from);
-    const newTo = normalizeMMDD(entry.to);
-    if (newFrom !== entry.from || newTo !== entry.to) {
-      console.log(`[SETTINGS] Migrating dates for "${entry.name}": ${entry.from}→${newFrom}, ${entry.to}→${newTo}`);
-      entry.from = newFrom;
-      entry.to = newTo;
-      changed = true;
-    }
-  }
-  if (changed) {
-    saveConfig().catch(err =>
-      console.error('[SETTINGS] Date migration save failed:', err.message)
-    );
-  }
-}
-
-/**
- * Migrate permanent and auto categories from plain string arrays to
- * object arrays with slug, description, and translations.
- * Backward-compat: stored Airtable config may still have the old format.
- */
-function migrateCategoryObjects() {
-  const sc = config.storefrontCategories;
-  if (!sc) return;
-  const emptyTranslations = { en: { title: '', description: '' }, pl: { title: '', description: '' }, ru: { title: '', description: '' }, uk: { title: '', description: '' } };
-  let changed = false;
-
-  // Permanent: string[] → object[]
-  if (Array.isArray(sc.permanent) && sc.permanent.length > 0 && typeof sc.permanent[0] === 'string') {
-    sc.permanent = sc.permanent.map(name => ({
-      name,
-      slug: name.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/-$/, ''),
-      description: '',
-      translations: JSON.parse(JSON.stringify(emptyTranslations)),
-    }));
-    changed = true;
-    console.log('[SETTINGS] Migrated permanent categories to object format');
-  }
-
-  // Auto: missing or string[]
-  if (!sc.auto) {
-    sc.auto = DEFAULTS.storefrontCategories.auto;
-    changed = true;
-  } else if (Array.isArray(sc.auto) && sc.auto.length > 0 && typeof sc.auto[0] === 'string') {
-    sc.auto = sc.auto.map(name => ({
-      name,
-      slug: name.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/-$/, ''),
-      description: '',
-      translations: JSON.parse(JSON.stringify(emptyTranslations)),
-    }));
-    changed = true;
-    console.log('[SETTINGS] Migrated auto categories to object format');
-  }
-
-  if (changed) {
-    saveConfig().catch(err =>
-      console.error('[SETTINGS] Category migration save failed:', err.message)
-    );
-  }
-}
-
-/**
- * Backfill default translations for auto categories (e.g. "Available Today")
- * when the stored config has empty title/description strings. The Wix site's
- * Velo code resolves menu labels by slug and falls through to the hardcoded
- * English `cat.name` when translations are empty — which is why the menu only
- * rendered in the English language. Filling in real EN/PL/RU/UK strings lets
- * the menu label translate automatically across all languages.
- */
-function migrateAutoCategoryTranslations() {
-  const sc = config.storefrontCategories;
-  if (!sc?.auto || !Array.isArray(sc.auto)) return;
-  let changed = false;
-
-  for (const stored of sc.auto) {
-    if (!stored || typeof stored !== 'object') continue;
-    const defaultEntry = (DEFAULTS.storefrontCategories.auto || [])
-      .find(d => d.slug === stored.slug);
-    if (!defaultEntry) continue;
-    if (!stored.translations) stored.translations = {};
-
-    for (const lang of ['en', 'pl', 'ru', 'uk']) {
-      const storedLang = stored.translations[lang] || {};
-      const defaultLang = defaultEntry.translations?.[lang] || {};
-      if (!storedLang.title && defaultLang.title) {
-        storedLang.title = defaultLang.title;
-        changed = true;
-      }
-      if (!storedLang.description && defaultLang.description) {
-        storedLang.description = defaultLang.description;
-        changed = true;
-      }
-      stored.translations[lang] = storedLang;
-    }
-
-    if (!stored.description && defaultEntry.description) {
-      stored.description = defaultEntry.description;
-      changed = true;
-    }
-  }
-
-  if (changed) {
-    console.log('[SETTINGS] Backfilled auto-category translations');
-    saveConfig().catch(err =>
-      console.error('[SETTINGS] Auto-category translation backfill save failed:', err.message)
-    );
-  }
-}
-
-/**
- * Migrate floristRates from flat format { name: number } to
- * per-type format { name: { type: number } }.
- * Old single rate becomes the first rateType (Standard).
- */
-function migrateFloristRates() {
-  const rates = config.floristRates;
-  if (!rates || typeof rates !== 'object') return;
-  let changed = false;
-  const defaultType = (config.rateTypes && config.rateTypes[0]) || 'Standard';
-  for (const name of Object.keys(rates)) {
-    if (typeof rates[name] === 'number') {
-      rates[name] = { [defaultType]: rates[name] };
-      changed = true;
-    }
-  }
-  if (changed) {
-    console.log('[SETTINGS] Migrated floristRates to per-type format');
-    saveConfig().catch(err =>
-      console.error('[SETTINGS] FloristRates migration save failed:', err.message)
-    );
-  }
-}
-
-// Load config on startup
-loadConfig();
-
-// ── Available Today cutoff reminder ─────────────────────────
-// Checks every minute. Once per day, after cutoff, if there are still
-// active Available Today products, sends a Telegram reminder to the owner.
-// Dedup date persisted in App Config so restarts don't re-fire the reminder.
-setInterval(async () => {
-  try {
-    if (!configLoaded) return;
-    const cutoff = config.availableTodayCutoff || '18:00';
-    const tz = config.availableTodayTimezone || 'Europe/Warsaw';
-    const now = new Date();
-    const timeStr = new Intl.DateTimeFormat('en-GB', {
-      hour: '2-digit', minute: '2-digit', hour12: false, timeZone: tz,
-    }).format(now);
-    const todayStr = new Intl.DateTimeFormat('en-CA', { timeZone: tz }).format(now);
-
-    if (timeStr < cutoff) return; // not past cutoff yet
-    if (config.cutoffReminderLastDate === todayStr) return; // already sent today
-
-    // Check if any active products with lead time 0 exist
-    const allActiveRows = await productConfigRepo.list({ activeOnly: true });
-    const zeroLeadRows = allActiveRows.filter(r => r['Lead Time Days'] === 0);
-
-    if (zeroLeadRows.length > 0) {
-      await sendAlert(
-        `⏰ Available Today reminder\n\n`
-        + `It's past ${cutoff} — you still have products marked as Available Today.\n`
-        + `Deactivate them in the dashboard if they're no longer available.`
-      );
-      config.cutoffReminderLastDate = todayStr;
-      saveConfig().catch(err => console.error('[SETTINGS] Failed to persist reminder date:', err.message));
-      console.log('[SETTINGS] Cutoff reminder sent');
-    }
-  } catch (err) {
-    console.error('[SETTINGS] Cutoff reminder error:', err.message);
-  }
-}, 60_000); // check every minute
-
-// ── Daily settings (auto-reset at midnight) ─────────────────
-const daily = {
-  driverOfDay:  null,
-  _lastSetDate: null,
-};
-
-// Build driver names from env vars (same pattern as auth.js)
-const driverNames = Object.entries(process.env)
-  .filter(([key]) => key.startsWith('PIN_DRIVER_'))
-  .map(([key]) =>
-    key.replace('PIN_DRIVER_', '').charAt(0).toUpperCase()
-    + key.replace('PIN_DRIVER_', '').slice(1).toLowerCase()
-  );
-
-function autoClearIfNewDay() {
-  const today = new Date().toISOString().split('T')[0];
-  if (daily._lastSetDate && daily._lastSetDate !== today) {
-    daily.driverOfDay = null;
-    daily._lastSetDate = null;
-  }
-}
 
 // ── GET /api/settings — read all settings + config ──
 router.get('/', authorize('orders'), (req, res) => {
   autoClearIfNewDay();
   const backupName = getBackupDriverName();
+  const daily = getDailyState();
+  const config = getAllConfig();
   const resolvedDrivers = [...new Set([...driverNames, ...config.extraDrivers])]
     .map(name => name === 'Backup' && backupName ? backupName : name);
   res.json({
@@ -371,12 +36,10 @@ router.get('/', authorize('orders'), (req, res) => {
 
 // ── PUT /api/settings/driver-of-day ──
 // When a driver-of-day is set, auto-assign them to all today's unassigned deliveries.
-// Like a shift supervisor assigning the day's driver to all open dispatches at once.
 router.put('/driver-of-day', authorize('admin'), async (req, res, next) => {
   try {
     const { driverName } = req.body;
-    daily.driverOfDay = driverName || null;
-    daily._lastSetDate = driverName ? new Date().toISOString().split('T')[0] : null;
+    setDailyDriver(driverName);
 
     let assignedCount = 0;
     if (driverName) {
@@ -391,7 +54,7 @@ router.put('/driver-of-day', authorize('admin'), async (req, res, next) => {
       }
     }
 
-    res.json({ driverOfDay: daily.driverOfDay, assignedCount });
+    res.json({ driverOfDay: getDriverOfDay(), assignedCount });
   } catch (err) {
     next(err);
   }
@@ -404,118 +67,23 @@ router.put('/backup-driver', authorize('admin'), (req, res) => {
   res.json({ backupDriverName: getBackupDriverName() });
 });
 
-// ── PUT /api/settings/config — update + persist to Airtable ──
+// ── PUT /api/settings/config — update + persist ──
 router.put('/config', authorize('admin'), async (req, res) => {
-  const allowed = Object.keys(config);
-  const updates = req.body;
-
-  for (const key of Object.keys(updates)) {
-    if (allowed.includes(key)) {
-      config[key] = updates[key];
-    }
-  }
-
-  // Persist to Airtable (fire-and-forget — don't block the response)
-  saveConfig().catch(err =>
-    console.error('[SETTINGS] Background save failed:', err.message)
-  );
-
-  res.json({ config });
+  updateConfigBulk(req.body);
+  res.json({ config: getAllConfig() });
 });
 
 // ── GET /api/settings/lists ──
 router.get('/lists', authorize('orders'), (req, res) => {
   res.json({
-    suppliers:      config.suppliers,
-    categories:     config.stockCategories,
-    paymentMethods: config.paymentMethods,
-    orderSources:   config.orderSources,
-    floristNames:   config.floristNames,
-    rateTypes:      config.rateTypes,
-    floristRates:   config.floristRates,
+    suppliers:      getConfig('suppliers'),
+    categories:     getConfig('stockCategories'),
+    paymentMethods: getConfig('paymentMethods'),
+    orderSources:   getConfig('orderSources'),
+    floristNames:   getConfig('floristNames'),
+    rateTypes:      getConfig('rateTypes'),
+    floristRates:   getConfig('floristRates'),
   });
 });
-
-// ── Exported getters for use by other modules ──
-
-export function getDriverOfDay() {
-  autoClearIfNewDay();
-  return daily.driverOfDay;
-}
-
-export function getConfig(key) {
-  return config[key];
-}
-
-export function updateConfig(key, value) {
-  config[key] = value;
-  saveConfig().catch(err =>
-    console.error('[SETTINGS] Background save failed:', err.message)
-  );
-}
-
-/**
- * Generate next App Order ID in YYYYMM-NNN format.
- * Counter stored in app_config Postgres table under key 'orderCounters'.
- * Like a sequential work order numbering system — each month resets the sequence.
- */
-export async function generateOrderId() {
-  const now      = new Date();
-  const monthKey = `${now.getFullYear()}${String(now.getMonth() + 1).padStart(2, '0')}`;
-  try {
-    return await appConfigRepo.nextOrderId(monthKey);
-  } catch (err) {
-    console.error('[ORDER-ID] Counter error:', err.message);
-    return `${monthKey}-T${Date.now().toString().slice(-5)}`;
-  }
-}
-
-/**
- * Check if current time in the configured timezone is past the cutoff.
- * Used for reminder logic — NOT for auto-hiding (owner controls visibility).
- */
-export function isPastCutoff() {
-  const cutoff = config.availableTodayCutoff || '18:00';
-  const tz = config.availableTodayTimezone || 'Europe/Warsaw';
-  const now = new Date();
-  const timeStr = new Intl.DateTimeFormat('en-GB', {
-    hour: '2-digit', minute: '2-digit',
-    hour12: false, timeZone: tz,
-  }).format(now);
-  return timeStr >= cutoff;
-}
-
-/**
- * Returns the currently active seasonal category based on config rules.
- * Priority: manual override > auto-schedule by date > null (none active).
- */
-export function getActiveSeasonalCategory() {
-  const sc = config.storefrontCategories;
-
-  if (sc.manualOverride) {
-    const forced = sc.seasonal.find(s => s.slug === sc.manualOverride);
-    if (forced) return {
-      name: forced.name, slug: forced.slug,
-      description: forced.description || '',
-      translations: forced.translations || {},
-    };
-  }
-
-  if (sc.autoSchedule) {
-    const now = new Date();
-    const mmdd = `${String(now.getMonth() + 1).padStart(2, '0')}-${String(now.getDate()).padStart(2, '0')}`;
-    for (const s of sc.seasonal) {
-      if (mmdd >= s.from && mmdd <= s.to) {
-        return {
-          name: s.name, slug: s.slug,
-          description: s.description || '',
-          translations: s.translations || {},
-        };
-      }
-    }
-  }
-
-  return null;
-}
 
 export default router;
